web应用/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
numpy==1.24.3
tables==3.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
